address minor PR feedbacks
