+++ express/schemas/component_spec.json
@@ -16,7 +16,7 @@
       "type": "string",
       "description": "Docker image for the component"
     },
-    "input_subsets": {
+     "input_subsets": {
       "type": "object",
       "description": "Input subsets for the component",
       "additionalProperties": {
@@ -25,8 +25,17 @@
           "fields": {
             "type": "object",
             "additionalProperties": {
-              "type": "string",
-              "$ref": "#/definitions/data_type"
+              "type": "object",
+              "properties": {
+                "type": {
+                  "type": "string",
+                  "$ref": "#/definitions/data_type"
+                },
+                "description": {
+                  "type": "string"
+                }
+              },
+              "required": ["type"]
             }
           }
         },
@@ -34,7 +43,7 @@
         "additionalProperties": false
       }
     },
-    "output_subsets": {
+     "output_subsets": {
       "type": "object",
       "description": "Output subsets for the component",
       "additionalProperties": {
@@ -43,8 +52,17 @@
           "fields": {
             "type": "object",
             "additionalProperties": {
-              "type": "string",
-              "$ref": "#/definitions/data_type"
+              "type": "object",
+              "properties": {
+                "type": {
+                  "type": "string",
+                  "$ref": "#/definitions/data_type"
+                },
+                "description": {
+                  "type": "string"
+                }
+              },
+              "required": ["type"]
             }
           }
         },
@@ -59,7 +77,7 @@
           "properties": {
             "type": {
               "type": "string",
-              "enum": ["list", "str"]
+              "enum": ["list", "str", "int"]
             },
             "description": {
               "type": "string"
